authorization for all route

+++ components/services/controller.js
@@ -1,30 +1,36 @@
 const _ = require('lodash')
 const BaseController = require('../base/controller')
 const Service = require('./service')
-const { SUCCESS, ERROR, STATUS, ROLES } = require('../../libs/constants')
+const { responses: RESPONSES, PRIVILAGES } = require('../../libs/constant')
 
 class Controller extends BaseController {
-  constructor () {
+  constructor() {
     super()
-    this.userCanGetListUpdateCreate = [ROLES.SUPER_ADMIN]
   }
 
-  async save (req, res) {
-    // if (_.includes(this.userCanGetListUpdateCreate, req.userData.role)) {
-      if (true) {
+  async save(req, res) {
+    const ROUTE_PRIVILAGE = [PRIVILAGES.USER.VALUE, PRIVILAGES.SUPER_ADMIN.VALUE]
 
-      const data = await new Service().save(req)
-      if (!_.isUndefined(data) && data.id) {
-        this.sendResponse(req, res, SUCCESS.CODE, { message: 'server_success.created_service' })
-      } else {
-        this.sendResponse(req, res, ERROR.CODE, { message: 'server_error.try_again' })
-      }
-    } else {
-      this.sendResponse(req, res, ERROR.CODE, { message: 'server_error.unauthorized' })
+    if (!_.includes(ROUTE_PRIVILAGE, req.user.userType)) {
+      return this.sendResponse(req, res, RESPONSES.UNAUTHORIZED_REQUEST, { message: 'unauthorized resources' })
     }
+
+
+    const data = await new Service().save(req)
+    if (!_.isUndefined(data) && data.id) {
+      this.sendResponse(req, res, RESPONSES.SUCCESS, { message: 'server_success.created_service' })
+    } else {
+      this.sendResponse(req, res, RESPONSES.ERROR, { message: 'server_error.try_again' })
+    }
+
   }
 
-  async list (req, res) {
+  async list(req, res) {
+    const ROUTE_PRIVILAGE = [PRIVILAGES.USER.VALUE, PRIVILAGES.SUPER_ADMIN.VALUE]
+
+    if (!_.includes(ROUTE_PRIVILAGE, req.user.userType)) {
+      return this.sendResponse(req, res, RESPONSES.UNAUTHORIZED_REQUEST, { message: 'unauthorized resources' })
+    }
     let skipPaging = false
     if (req.query && req.query.skipPagination && req.query.skipPagination === 'true') {
       skipPaging = true
@@ -32,31 +38,40 @@ class Controller extends BaseController {
     const data = await new Service().list(req, skipPaging)
 
     if (!_.isUndefined(data) && data.length > 0) {
-      this.sendResponse(req, res, SUCCESS.CODE, { data, count: data.length })
+      this.sendResponse(req, res, RESPONSES.SUCCESS, { data, count: data.length })
     } else {
-      this.sendResponse(req, res, ERROR.CODE, { message: 'server_error.no_records_found' })
+      this.sendResponse(req, res, RESPONSES.ERROR, { message: 'server_error.no_records_found' })
     }
   }
 
-  async update (req, res){
+  async update(req, res) {
+    const ROUTE_PRIVILAGE = [PRIVILAGES.USER.VALUE, PRIVILAGES.SUPER_ADMIN.VALUE]
 
+    if (!_.includes(ROUTE_PRIVILAGE, req.user.userType)) {
+      return this.sendResponse(req, res, RESPONSES.UNAUTHORIZED_REQUEST, { message: 'unauthorized resources' })
+    }
     const data = await new Service().update(req)
-    if(!_.isUndefined(data)){
-        this.sendResponse(req, res, SUCCESS.CODE, {data})
-    }else{
-        this.sendResponse(req, res, ERROR.CODE, {message: 'server_error.try_again'})
+    if (!_.isUndefined(data)) {
+      this.sendResponse(req, res, RESPONSES.SUCCESS, { data })
+    } else {
+      this.sendResponse(req, res, RESPONSES.ERROR, { message: 'server_error.try_again' })
     }
   }
 
-  async delete (req, res){
+  async delete(req, res) {
+    const ROUTE_PRIVILAGE = [PRIVILAGES.USER.VALUE, PRIVILAGES.SUPER_ADMIN.VALUE]
+
+    if (!_.includes(ROUTE_PRIVILAGE, req.user.userType)) {
+      return this.sendResponse(req, res, RESPONSES.UNAUTHORIZED_REQUEST, { message: 'unauthorized resources' })
+    }
 
     const data = await new Service().delete(req)
-    if(!_.isUndefined(data)){
-        this.sendResponse(req, res, SUCCESS.CODE, {data})
-    }else{
-        this.sendResponse(req, res, ERROR.CODE, {message: 'server_error.try_again'})
+    if (!_.isUndefined(data)) {
+      this.sendResponse(req, res, RESPONSES.SUCCESS, { data })
+    } else {
+      this.sendResponse(req, res, RESPONSES.ERROR, { message: 'server_error.try_again' })
     }
-}
+  }
 }
